finished block pinching and refactored soem code
