+++ python/pygfdm/CMakeLists.txt
@@ -23,6 +23,7 @@
 GR_PYTHON_INSTALL(
     FILES
     filters.py
+    cyclic_prefix.py
     __init__.py
     DESTINATION ${GR_PYTHON_DIR}/gfdm/pygfdm
 )
