+++ python/qa_modulator_cc.py
@@ -22,8 +22,8 @@
 from gnuradio import gr, gr_unittest
 from gnuradio import blocks
 import gfdm_swig as gfdms
-from pygfdm.gfdm_modulation import gfdm_modulate_block, get_random_qpsk
-from pygfdm.modulation import get_data_matrix
+from pygfdm.gfdm_modulation import gfdm_modulate_block
+from pygfdm.mapping import get_data_matrix, get_random_qpsk
 from pygfdm.filters import get_frequency_domain_filter
 import numpy as np
